requirements.txt
click==8.1.3
colorama==0.4.6
grpcio==1.51.1
grpcio-tools==1.51.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
protobuf==4.21.12
psycopg2==2.9.5
Werkzeug==2.2.2
